refactor: extract domain layer into @costwise/domain
Verification gates output:

1. pnpm build:
@costwise/db:build: $ tsc --noEmit
@costwise/shared:build: $ tsc --noEmit
@costwise/domain:build: $ tsc --noEmit
web:build: $ next build
Compiled successfully in 1000ms
Linting and checking validity of types ...
Generating static pages (15/15)
Tasks: 4 successful, 4 total

2. pnpm test:
Test Suites: 6 passed, 6 total
Tests: 28 passed, 28 total

3. pnpm lint:
Tasks: 1 successful, 1 total (0 errors)

4. grep -rn "next/" packages/:
(empty, exit code 1)

5. grep web moved module imports:
Only web-local client helpers in services.ts remain.
