+++ apps/web/src/app/(user)/recipes/create/page.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import RecipeForm from '@/app/components/recipes/recipeForm/recipeForm';
-import { IngredientService } from '@/app/services/ingredientService';
+import { IngredientService } from '@costwise/domain/services/ingredientService';
 import { RecipeIngredients } from '@costwise/shared/recipe';
 import { auth } from '@/auth';
 import { redirect } from 'next/navigation';
-import { Metadata } from '@/types/specialTypes';
+import { Metadata } from '@costwise/domain/types/specialTypes';
 
 const Page = async () => {
   const session = await auth();
