+++ apps/web/src/app/api/ingredients/[id]/route.ts
@@ -13,10 +13,10 @@
  */
 import { NextRequest } from "next/server";
 import { sendSuccess } from "../../utils/responses";
-import { IngredientService } from "@/app/services/ingredientService";
+import { IngredientService } from "@costwise/domain/services/ingredientService";
 import { Ingredient } from "@costwise/shared/recipe";
 import { auth } from "@/auth";
-import { AuthenticationError } from "@/app/utils/errors";
+import { AuthenticationError } from "@costwise/domain/utils/errors";
 import { errorHandler } from "@/app/utils/errorHandler";
 
 export const PATCH = async (req: NextRequest) => {
